+++ cs/TagsCloudVisualization/CircularCloudLayouter.cs
@@ -9,30 +9,27 @@ namespace TagsCloudVisualization
     {
         public Cloud Cloud { get; }
         private readonly IEnumerator<Point> spiralEnumerator;
-        private Size currentRectangleSize;
 
         public CircularCloudLayouter(Point center)
         {
             Cloud = new Cloud(center);
-            spiralEnumerator = SpiralGenerator.GetSpiral(center, 5, Math.PI / 16).GetEnumerator();
+            spiralEnumerator = SpiralGenerator.GetSpiral(center, 0.5, Math.PI / 16).GetEnumerator();
         }
 
         public Rectangle PutNextRectangle(Size rectangleSize)
         {
-            currentRectangleSize = rectangleSize;
-
-            var position = GetRectanglePosition();
+            var position = GetRectangleLocation(rectangleSize);
             var rectangle = new Rectangle(position, rectangleSize);
             Cloud.Rectangles.Add(rectangle);
 
             return rectangle;
         }
 
-        private Point GetRectanglePosition()
+        private Point GetRectangleLocation(Size rectangleSize)
         {
             Point location;
 
-            while (!TryGetNextLocation(out location))
+            while (!TryGetNextLocation(rectangleSize, out location))
             {
                 //Should be empty
             }
@@ -40,23 +37,22 @@ namespace TagsCloudVisualization
             return location;
         }
 
-        private bool TryGetNextLocation(out Point location)
+        private bool TryGetNextLocation(Size rectangleSize, out Point upperLeftCorner)
         {
             spiralEnumerator.MoveNext();
-            var upperLeftCorner = spiralEnumerator.Current;
-            
-            location = GetCenterPosition(upperLeftCorner);
-            var rectangle = new Rectangle(location, currentRectangleSize);
+            var center = spiralEnumerator.Current;
 
+            upperLeftCorner = GetUpperLeftCornerPosition(rectangleSize, center);
+            var rectangle = new Rectangle(upperLeftCorner, rectangleSize);
             return Cloud.Rectangles.All(r => !r.IntersectsWith(rectangle));
         }
 
-        private Point GetCenterPosition(Point corner)
+        protected Point GetUpperLeftCornerPosition(Size rectangleSize, Point center)
         {
-            var xOffset = currentRectangleSize.Width / 2;
-            var yOffset = currentRectangleSize.Height / 2;
+            var xOffset = rectangleSize.Width / 2;
+            var yOffset = rectangleSize.Height / 2;
 
-            return new Point(corner.X - xOffset, corner.Y - yOffset);
+            return new Point(center.X - xOffset, center.Y - yOffset);
         }
     }
 }
